move controllers to V1 namespace
