+++ OutOfSchool/OutOfSchool.WebApi.Tests/Controllers/ApplicationControllerTests.cs
@@ -12,6 +12,7 @@ using NUnit.Framework;
 using OutOfSchool.Services.Enums;
 using OutOfSchool.WebApi.ApiModels;
 using OutOfSchool.WebApi.Controllers;
+using OutOfSchool.WebApi.Controllers.V1;
 using OutOfSchool.WebApi.Models;
 using OutOfSchool.WebApi.Services;
