+++ OutOfSchool/OutOfSchool.WebApi.Tests/Controllers/CityControllerTests.cs
@@ -7,6 +7,7 @@ using Microsoft.Extensions.Localization;
 using Moq;
 using NUnit.Framework;
 using OutOfSchool.WebApi.Controllers;
+using OutOfSchool.WebApi.Controllers.V1;
 using OutOfSchool.WebApi.Models;
 using OutOfSchool.WebApi.Services;
