+++ OutOfSchool/OutOfSchool.WebApi.Tests/Controllers/DepartmentControllerTests.cs
@@ -11,6 +11,7 @@ using NUnit.Framework;
 using OutOfSchool.Services.Enums;
 using OutOfSchool.Services.Models;
 using OutOfSchool.WebApi.Controllers;
+using OutOfSchool.WebApi.Controllers.V1;
 using OutOfSchool.WebApi.Models;
 using OutOfSchool.WebApi.Services;
