+++ OutOfSchool/OutOfSchool.WebApi.Tests/Controllers/ParentControllerTests.cs
@@ -10,6 +10,7 @@ using Moq;
 using NUnit.Framework;
 using OutOfSchool.Services.Enums;
 using OutOfSchool.WebApi.Controllers;
+using OutOfSchool.WebApi.Controllers.V1;
 using OutOfSchool.WebApi.Models;
 using OutOfSchool.WebApi.Services;
