+++ OutOfSchool/OutOfSchool.WebApi.Tests/Controllers/ProviderControllerTests.cs
@@ -15,6 +15,7 @@ using NUnit.Framework;
 using OutOfSchool.Services.Enums;
 using OutOfSchool.Tests.Common;
 using OutOfSchool.WebApi.Controllers;
+using OutOfSchool.WebApi.Controllers.V1;
 using OutOfSchool.WebApi.Models;
 using OutOfSchool.WebApi.Services;
