+++ OutOfSchool/OutOfSchool.WebApi.Tests/Controllers/RatingControllerTests.cs
@@ -9,6 +9,7 @@ using NUnit.Framework;
 using OutOfSchool.ElasticsearchData.Models;
 using OutOfSchool.Services.Enums;
 using OutOfSchool.WebApi.Controllers;
+using OutOfSchool.WebApi.Controllers.V1;
 using OutOfSchool.WebApi.Models;
 using OutOfSchool.WebApi.Services;
