+++ OutOfSchool/OutOfSchool.WebApi/Controllers/V1/ChildController.cs
@@ -8,13 +8,14 @@ using OutOfSchool.WebApi.Extensions;
 using OutOfSchool.WebApi.Models;
 using OutOfSchool.WebApi.Services;
 
-namespace OutOfSchool.WebApi.Controllers
+namespace OutOfSchool.WebApi.Controllers.V1
 {
     /// <summary>
     /// Controller with CRUD operations for a Child entity.
     /// </summary>
     [ApiController]
-    [Route("[controller]/[action]")]
+    [ApiVersion("1.0")]
+    [Route("api/v{version:apiVersion}/[controller]/[action]")]
     [Authorize(AuthenticationSchemes = "Bearer")]
     public class ChildController : ControllerBase
     {
