+++ OutOfSchool/OutOfSchool.WebApi/Controllers/V1/ElasticsearchWorkshopController.cs
@@ -4,10 +4,11 @@ using Microsoft.AspNetCore.Mvc;
 using OutOfSchool.ElasticsearchData.Models;
 using OutOfSchool.WebApi.Services;
 
-namespace OutOfSchool.WebApi.Controllers
+namespace OutOfSchool.WebApi.Controllers.V1
 {
     [ApiController]
-    [Route("[controller]/[action]")]
+    [ApiVersion("1.0")]
+    [Route("api/v{version:apiVersion}/[controller]/[action]")]
     [Authorize(AuthenticationSchemes = "Bearer")]
     [Authorize(Roles = "admin")]
     public class ElasticsearchWorkshopController : ControllerBase
