+++ OutOfSchool/OutOfSchool.WebApi/Controllers/V1/SocialGroupController.cs
@@ -1,5 +1,4 @@
-using System;
-using System.Linq;
+﻿using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
@@ -9,56 +8,56 @@ using OutOfSchool.WebApi.Extensions;
 using OutOfSchool.WebApi.Models;
 using OutOfSchool.WebApi.Services;
 
-namespace OutOfSchool.WebApi.Controllers
+namespace OutOfSchool.WebApi.Controllers.V1
 {
-    /// <summary>
-    /// Controller with CRUD operations for a Teacher entity.
-    /// </summary>
     [ApiController]
-    [Route("[controller]/[action]")]
-    public class TeacherController : ControllerBase
+    [ApiVersion("1.0")]
+    [Route("api/v{version:apiVersion}/[controller]/[action]")]
+    [Authorize(AuthenticationSchemes = "Bearer")]
+    public class SocialGroupController : ControllerBase
     {
-        private readonly ITeacherService service;
+        private readonly ISocialGroupService service;
         private readonly IStringLocalizer<SharedResource> localizer;
 
         /// <summary>
-        /// Initializes a new instance of the <see cref="TeacherController"/> class.
+        /// Initializes a new instance of the <see cref="SocialGroupController"/> class.
         /// </summary>
-        /// <param name="service">Service for Teacher model.</param>
+        /// <param name="service">Service for SocialGroup model.</param>
         /// <param name="localizer">Localizer.</param>
-        public TeacherController(ITeacherService service, IStringLocalizer<SharedResource> localizer)
+        public SocialGroupController(ISocialGroupService service, IStringLocalizer<SharedResource> localizer)
         {
-            this.localizer = localizer;
             this.service = service;
+            this.localizer = localizer;
         }
 
         /// <summary>
-        /// Get all teachers from the database.
+        /// Get all Social Groups from the database.
         /// </summary>
-        /// <returns>List of teachers.</returns>
+        /// <returns>List of all Social Groups.</returns>
         [ProducesResponseType(StatusCodes.Status200OK)]
         [ProducesResponseType(StatusCodes.Status204NoContent)]
+        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
         [ProducesResponseType(StatusCodes.Status500InternalServerError)]
         [HttpGet]
         public async Task<IActionResult> Get()
         {
-            var teachers = await service.GetAll().ConfigureAwait(false);
+            var socialGroup = await service.GetAll().ConfigureAwait(false);
 
-            if (!teachers.Any())
+            if (!socialGroup.Any())
             {
                 return NoContent();
             }
 
-            return Ok(teachers);
+            return Ok(socialGroup);
         }
 
         /// <summary>
-        /// Get teacher by it's id.
+        /// Get Social Group by it's id.
         /// </summary>
-        /// <param name="id">Teacher's id.</param>
-        /// <returns>Teacher.</returns>
+        /// <param name="id">Social Group id.</param>
+        /// <returns>Social Group.</returns>
         [ProducesResponseType(StatusCodes.Status200OK)]
-        [ProducesResponseType(StatusCodes.Status204NoContent)]
+        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
         [ProducesResponseType(StatusCodes.Status500InternalServerError)]
         [HttpGet("{id}")]
         public async Task<IActionResult> GetById(long id)
@@ -69,60 +68,50 @@ namespace OutOfSchool.WebApi.Controllers
         }
 
         /// <summary>
-        /// Add a new teacher to the database.
+        /// Add a new Social Group to the database.
         /// </summary>
-        /// <param name="dto">Entity to add.</param>
+        /// <param name="dto">Social Group entity to add.</param>
         /// <returns>A <see cref="Task{TResult}"/> representing the result of the asynchronous operation.</returns>
-        [Authorize(Roles = "provider,admin")]
-        [Authorize(AuthenticationSchemes = "Bearer")]
+        [Authorize(Roles = "admin")]
         [ProducesResponseType(StatusCodes.Status201Created)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
         [ProducesResponseType(StatusCodes.Status500InternalServerError)]
         [HttpPost]
-        public async Task<IActionResult> Create(TeacherDTO dto)
+        public async Task<IActionResult> Create(SocialGroupDto dto)
         {
-            if (!ModelState.IsValid)
-            {
-                return BadRequest(ModelState);
-            }
-
-            var teacher = await service.Create(dto).ConfigureAwait(false);
+            var socialGroup = await service.Create(dto).ConfigureAwait(false);
 
             return CreatedAtAction(
                 nameof(GetById),
-                new { id = teacher.Id, },
-                teacher);
+                new { id = socialGroup.Id, },
+                socialGroup);
         }
 
         /// <summary>
-        /// Update info about a specific teacher in the database.
+        /// Update info about a Social Group in the database.
         /// </summary>
-        /// <param name="dto">Teacher to update.</param>
-        /// <returns>Teacher.</returns>
-        [Authorize(Roles = "provider,admin")]
-        [Authorize(AuthenticationSchemes = "Bearer")]
+        /// <param name="dto">Social Group to update.</param>
+        /// <returns>Social Group.</returns>
+        [Authorize(Roles = "admin")]
         [ProducesResponseType(StatusCodes.Status200OK)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
         [ProducesResponseType(StatusCodes.Status500InternalServerError)]
         [HttpPut]
-        public async Task<IActionResult> Update(TeacherDTO dto)
+        public async Task<IActionResult> Update(SocialGroupDto dto)
         {
-            if (!ModelState.IsValid)
-            {
-                return BadRequest(ModelState);
-            }
-
             return Ok(await service.Update(dto).ConfigureAwait(false));
         }
 
         /// <summary>
-        /// Delete a specific Teacher entity from the database.
+        /// Delete a specific Social Group from the database.
         /// </summary>
-        /// <param name="id">Teacher's id.</param>
+        /// <param name="id">Social Group id.</param>
         /// <returns>A <see cref="Task{TResult}"/> representing the result of the asynchronous operation.</returns>
-        [Authorize(Roles = "provider,admin")]
-        [Authorize(AuthenticationSchemes = "Bearer")]
+        [Authorize(Roles = "admin")]
         [ProducesResponseType(StatusCodes.Status204NoContent)]
+        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
         [ProducesResponseType(StatusCodes.Status500InternalServerError)]
         [HttpDelete("{id}")]
         public async Task<IActionResult> Delete(long id)
