+++ OutOfSchool/OutOfSchool.WebApi/Controllers/V1/StatisticController.cs
@@ -7,12 +7,13 @@ using Microsoft.AspNetCore.Mvc;
 using OutOfSchool.WebApi.Models;
 using OutOfSchool.WebApi.Services;
 
-namespace OutOfSchool.WebApi.Controllers
+namespace OutOfSchool.WebApi.Controllers.V1
 {
     /// <summary>
     /// Controller with operations to get popular workshops and categories.
     /// </summary>
-    [Route("[controller]/[action]")]
+    [ApiVersion("1.0")]
+    [Route("api/v{version:apiVersion}/[controller]/[action]")]
     [ApiController]
     [Authorize(AuthenticationSchemes = "Bearer")]
